:recycle: refactor file paths
