+++ src/renderer/data/constants/device-styles.ts
