+++ src/renderer/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from './use-compiler'
 export * from './use-remove-tab'
+export * from './use-store-selectors'
